atomic: Rewrite origin to point to metalink
Basically Anaconda-in-ImageFactory is set up to pull from the
builder, so that's what ends up in the origin file.  But that's
obviously not what we want to ship to users.

ostree/rpm-ostree do not yet have a convenient command to change this,
so brute force it with sed.

Implements https://lists.fedoraproject.org/pipermail/cloud/2014-November/004570.html