fix ostree repo location
problem: fedora atomic host images are including an old
ostree tree (24.39 from 2016-06-15 09:57:04)

cause: The ostree repo specified in this file at:
https://kojipkgs.fedoraproject.org/compose/atomic/24/
was last updated at 2016-06-15 09:57:04.

Fedora Atomic 23 used the location:
https://kojipkgs.fedoraproject.org/mash/atomic/23/.

The ref at the similar
https://kojipkgs.fedoraproject.org/mash/atomic/24/
was last modifed 2016-07-24. This seems to be the correct
location. Or, if the switch from mash to compose is right,
then we need to ensure that the compose location is updated
/ kept up to date.