KDE: fix compose due to core deps needing make
So make is needed by openssl and compat-openssl10 which is turn is needed
by key things such as qt5-qtbase, mariadb and other core things that are
explictly needed in the KDE spin. I have no idea why the ARM images fail
and the live media does not because all arches have this dep.

Signed-off-by: Peter Robinson <pbrobinson@fedoraproject.org>