Drop fedora-live-minimization use from cinnamon
live-minimization is kinda old, and dates from the time we were
trying to fit in CD size. No-one cares much about that any more.
One of these exclusions (sane-backends) was contradictory for
Cinnamon lives and broke the compose since the recent dnf change
to fail compose on dependency issues (because the cinnamon comps
group explicitly includes sane-backends-drivers-scanners, but
this exclusion excludes one of its deps). To resolve that,
@grinnz is fine with just dropping the use of live-minimization
from the cinnamon images entirely, so this does that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>