Stop using fedora-live-minimization from Python Classroom
From Adam Williamson:

The Python Classroom spin images no longer build in recent Rawhide,
because of a dnf behaviour change that exposes dependency issues as
failures rather than hiding them by silently excluding packages. (This
change will likely get into F29 soon too).

The issue is this. fedora-live-python-classroom.ks includes
fedora-live-minimization.ks , which does this:

-sane-backends

but it also includes fedora-live-workstation.ks, which includes fedora-
workstation-common.ks, which does this:

@gnome-desktop

...and @gnome-desktop includes sane-backends-drivers-scanners and
libsane-hpaio, which both require sane-backends.

Until recently dnf was simply silently excluding sane-backends-drivers-
scanners and libsane-hpaio from the image, to 'resolve' this. Now it
fails on the problem, and we get to make a choice.

You basically have three options:

1. Just drop fedora-live-minimization.ks from fedora-live-python-
classroom.ks . It's pretty old and random at this point. It's basically
trying to drop printing and scanning stuff to save a bit of space. Do
you actually want printing and scanning not to work on your image?

2. Keep fedora-live-minimization.ks but explicitly add back `sane-
backends` to %packages in fedora-live-python-classroom.ks . This would
override the exclusion and fix the dep issue, and include the scanning
bits in the image.

3. Keep fedora-live-minimization.ks and add excludes for sane-backends-
drivers-scanners and libsane-hpaio to fedora-live-python-classroom.ks .
This would lose scanning support, and save a bit of space.

I went with option 1.