Revert "remove systemd package from fedora-container-base"
This reverts commit 57e13a1b8970c2e15d0c310aa871e4737781a23f.

I noticed that systemd-logind was causing problems in our
containers after a container rebuild of the f32 base image was
published about 5 days ago. I found that an issue[0] had been
filed about the problem, but that bug was reported against
Rawhide (f33). I then found that systemd had been removed from
the f32 base image after f32 was released.

This causes problems for users who expect stability out of Fedora
releases - changing the default package set means that a container
build that works one day can break the next. In our case, the
problem wasn't so much that systemd was removed, but that the unit
masks were removed. This causes some components of systemd to be
unable to function in a container, and puts a burden on users to
identify the problem.

https://pagure.io/releng/issue/9603

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1841139