fedora-live-kde.ks: Really enable initial-setup on firstboot
Previous change enabled it for the LiveISO directly instead of the final
system as this was meant for usage in RAW disk images and not in
LiveISOs.

This change should enable it for both, be a NOP for the LiveISO and
effectively enable it on the final system until we find a better
solution in Anaconda itself.

See discussion in https://github.com/rhinstaller/initial-setup/issues/136

Fixes: 8a3b18b fedora-live-kde.ks: Enable initial-setup on firstboot