🚮 Truncate /etc/resolv.conf from kickstart
During the kickstart of the cloud image, NetworkManager writes an
/etc/resolv.conf that contains `nameserver 192.168.122.1`. This causes
boot delays with cloud-init since it does some early boot DNS
redirection tests before talking to the cloud's metadata service. On
some clouds/architectures, this delay is 15 seconds or more.

Truncate the /etc/resolv.conf so it can be replaced properly by
NetworkManager and cloud-init on the first boot.

Signed-off-by: Major Hayden <major@redhat.com>