Disable automount on KDE live images (#2073708)
It's mounting non-removable devices and causing problems for
anaconda. Note, KDE will still show a notification with a
button for mounting the device, and it shows in the file manager
where clicking on it mounts it; the behaviour is about the same
as Fedora 35 (slightly better, as the notification shows on
subsequent plugs of the same stick, which it did not in F35).

Signed-off-by: Adam Williamson <awilliam@redhat.com>